api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.26.0
pgvector==0.2.4
numpy==1.26.3
asyncpg==0.29.0
jinja2==3.1.2
yfinance==0.2.36
PyJWT==2.8.0
bcrypt==4.1.2
redis==5.0.1
pandas==2.1.4
requests==2.31.0
textblob==0.17.1
prophet==1.1.5
scikit-learn==1.3.2
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2
faiss-cpu==1.7.4
google-generativeai==0.3.2
openai==1.7.2
anthropic==0.8.1
beautifulsoup4==4.12.2
lxml==5.1.0
cryptography==42.0.0
huggingface-hub==0.23.4
prometheus-client==0.19.0
